+++ src/ui/log_wizard.Designer.cs
@@ -104,6 +104,10 @@
             this.aboutToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
             this.whatupMenu = new System.Windows.Forms.ContextMenuStrip(this.components);
             this.refreshAddViewButtons = new System.Windows.Forms.Timer(this.components);
+            this.openLogSamplesToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.defaultSampleToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.bigLogToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.smallSampleLogToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.newViewMenu.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.toggleTopmost)).BeginInit();
             this.toggleMenu.SuspendLayout();
@@ -160,7 +164,7 @@
             this.newFilteredView.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.newFilteredView.ContextMenuStrip = this.newViewMenu;
             this.newFilteredView.Font = new System.Drawing.Font("Segoe UI", 7F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.newFilteredView.Location = new System.Drawing.Point(394, 29);
+            this.newFilteredView.Location = new System.Drawing.Point(392, 29);
             this.newFilteredView.Name = "newFilteredView";
             this.newFilteredView.Size = new System.Drawing.Size(18, 20);
             this.newFilteredView.TabIndex = 1;
@@ -195,7 +199,7 @@
             // 
             this.delFilteredView.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.delFilteredView.Font = new System.Drawing.Font("Segoe UI", 7F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.delFilteredView.Location = new System.Drawing.Point(412, 29);
+            this.delFilteredView.Location = new System.Drawing.Point(410, 29);
             this.delFilteredView.Name = "delFilteredView";
             this.delFilteredView.Size = new System.Drawing.Size(18, 20);
             this.delFilteredView.TabIndex = 2;
@@ -256,7 +260,7 @@
             this.synchronizedWithFullLog.Checked = true;
             this.synchronizedWithFullLog.CheckState = System.Windows.Forms.CheckState.Checked;
             this.synchronizedWithFullLog.Font = new System.Drawing.Font("Segoe UI", 7F);
-            this.synchronizedWithFullLog.Location = new System.Drawing.Point(286, 29);
+            this.synchronizedWithFullLog.Location = new System.Drawing.Point(284, 29);
             this.synchronizedWithFullLog.Name = "synchronizedWithFullLog";
             this.synchronizedWithFullLog.Size = new System.Drawing.Size(46, 20);
             this.synchronizedWithFullLog.TabIndex = 1;
@@ -272,7 +276,7 @@
             this.synchronizeWithExistingLogs.Checked = true;
             this.synchronizeWithExistingLogs.CheckState = System.Windows.Forms.CheckState.Checked;
             this.synchronizeWithExistingLogs.Font = new System.Drawing.Font("Segoe UI", 7F);
-            this.synchronizeWithExistingLogs.Location = new System.Drawing.Point(240, 29);
+            this.synchronizeWithExistingLogs.Location = new System.Drawing.Point(238, 29);
             this.synchronizeWithExistingLogs.Name = "synchronizeWithExistingLogs";
             this.synchronizeWithExistingLogs.Size = new System.Drawing.Size(46, 20);
             this.synchronizeWithExistingLogs.TabIndex = 3;
@@ -613,7 +617,7 @@
             this.splitDescription.Panel2.Controls.Add(this.description);
             this.splitDescription.Panel2MinSize = 100;
             this.splitDescription.Size = new System.Drawing.Size(974, 450);
-            this.splitDescription.SplitterDistance = 308;
+            this.splitDescription.SplitterDistance = 306;
             this.splitDescription.SplitterWidth = 6;
             this.splitDescription.TabIndex = 18;
             this.splitDescription.SplitterMoved += new System.Windows.Forms.SplitterEventHandler(this.splitDescription_SplitterMoved);
@@ -634,8 +638,8 @@
             this.filteredLeft.Panel1.Controls.Add(this.viewsTab);
             this.filteredLeft.Panel1MinSize = 100;
             this.filteredLeft.Panel2MinSize = 100;
-            this.filteredLeft.Size = new System.Drawing.Size(974, 308);
-            this.filteredLeft.SplitterDistance = 469;
+            this.filteredLeft.Size = new System.Drawing.Size(974, 306);
+            this.filteredLeft.SplitterDistance = 467;
             this.filteredLeft.SplitterWidth = 6;
             this.filteredLeft.TabIndex = 0;
             this.filteredLeft.SplitterMoved += new System.Windows.Forms.SplitterEventHandler(this.filteredLeft_SplitterMoved);
@@ -651,7 +655,7 @@
             this.viewsTab.Location = new System.Drawing.Point(0, 3);
             this.viewsTab.Name = "viewsTab";
             this.viewsTab.SelectedIndex = 0;
-            this.viewsTab.Size = new System.Drawing.Size(464, 306);
+            this.viewsTab.Size = new System.Drawing.Size(462, 304);
             this.viewsTab.TabIndex = 0;
             this.viewsTab.SelectedIndexChanged += new System.EventHandler(this.viewsTab_SelectedIndexChanged);
             this.viewsTab.DragDrop += new System.Windows.Forms.DragEventHandler(this.filteredViews_DragDrop);
@@ -663,7 +667,7 @@
             this.tabPage1.Location = new System.Drawing.Point(4, 24);
             this.tabPage1.Name = "tabPage1";
             this.tabPage1.Padding = new System.Windows.Forms.Padding(3);
-            this.tabPage1.Size = new System.Drawing.Size(456, 278);
+            this.tabPage1.Size = new System.Drawing.Size(454, 276);
             this.tabPage1.TabIndex = 0;
             this.tabPage1.Text = "View";
             this.tabPage1.UseVisualStyleBackColor = true;
@@ -678,7 +682,7 @@
             this.dropHere.ForeColor = System.Drawing.Color.FromArgb(((int)(((byte)(224)))), ((int)(((byte)(224)))), ((int)(((byte)(224)))));
             this.dropHere.Location = new System.Drawing.Point(5, 3);
             this.dropHere.Name = "dropHere";
-            this.dropHere.Size = new System.Drawing.Size(445, 278);
+            this.dropHere.Size = new System.Drawing.Size(443, 276);
             this.dropHere.TabIndex = 0;
             this.dropHere.Text = "Drop a Log file here\r\n to open";
             this.dropHere.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
@@ -691,7 +695,7 @@
             this.description.Font = new System.Drawing.Font("Segoe UI", 10F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.description.Location = new System.Drawing.Point(0, 0);
             this.description.Name = "description";
-            this.description.Size = new System.Drawing.Size(974, 136);
+            this.description.Size = new System.Drawing.Size(974, 138);
             this.description.TabIndex = 0;
             // 
             // refresh
@@ -871,8 +875,9 @@
             // whatupMenu
             // 
             this.whatupMenu.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
-            this.whatsupOpen,
             this.whatsupNew,
+            this.whatsupOpen,
+            this.openLogSamplesToolStripMenuItem,
             this.toolStripSeparator4,
             this.whatsupToggles,
             this.whatsupPreferences,
@@ -885,13 +890,44 @@
             this.hotkeysHelpToolStripMenuItem,
             this.aboutToolStripMenuItem1});
             this.whatupMenu.Name = "whatupMenu";
-            this.whatupMenu.Size = new System.Drawing.Size(205, 242);
+            this.whatupMenu.Size = new System.Drawing.Size(205, 286);
             // 
             // refreshAddViewButtons
             // 
             this.refreshAddViewButtons.Enabled = true;
             this.refreshAddViewButtons.Interval = 250;
             this.refreshAddViewButtons.Tick += new System.EventHandler(this.refreshAddViewButtons_Tick);
+            // 
+            // openLogSamplesToolStripMenuItem
+            // 
+            this.openLogSamplesToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.defaultSampleToolStripMenuItem,
+            this.bigLogToolStripMenuItem,
+            this.smallSampleLogToolStripMenuItem});
+            this.openLogSamplesToolStripMenuItem.Name = "openLogSamplesToolStripMenuItem";
+            this.openLogSamplesToolStripMenuItem.Size = new System.Drawing.Size(204, 22);
+            this.openLogSamplesToolStripMenuItem.Text = "Open Log Samples";
+            // 
+            // defaultSampleToolStripMenuItem
+            // 
+            this.defaultSampleToolStripMenuItem.Name = "defaultSampleToolStripMenuItem";
+            this.defaultSampleToolStripMenuItem.Size = new System.Drawing.Size(168, 22);
+            this.defaultSampleToolStripMenuItem.Text = "Default Sample";
+            this.defaultSampleToolStripMenuItem.Click += new System.EventHandler(this.defaultSampleToolStripMenuItem_Click);
+            // 
+            // bigLogToolStripMenuItem
+            // 
+            this.bigLogToolStripMenuItem.Name = "bigLogToolStripMenuItem";
+            this.bigLogToolStripMenuItem.Size = new System.Drawing.Size(168, 22);
+            this.bigLogToolStripMenuItem.Text = "Big Sample Log";
+            this.bigLogToolStripMenuItem.Click += new System.EventHandler(this.bigLogToolStripMenuItem_Click);
+            // 
+            // smallSampleLogToolStripMenuItem
+            // 
+            this.smallSampleLogToolStripMenuItem.Name = "smallSampleLogToolStripMenuItem";
+            this.smallSampleLogToolStripMenuItem.Size = new System.Drawing.Size(168, 22);
+            this.smallSampleLogToolStripMenuItem.Text = "Small Sample Log";
+            this.smallSampleLogToolStripMenuItem.Click += new System.EventHandler(this.smallSampleLogToolStripMenuItem_Click);
             // 
             // log_wizard
             // 
@@ -1021,6 +1057,10 @@
         private lw_common.ui.description_ctrl description;
         private System.Windows.Forms.ToolStripMenuItem exportCurrentViewToCSVToolStripMenuItem;
         private lw_common.ui.categories_ctrl categories;
+        private System.Windows.Forms.ToolStripMenuItem openLogSamplesToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem defaultSampleToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem bigLogToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem smallSampleLogToolStripMenuItem;
     }
 }